move shared maven utilities to a dedicated class
